+++ src/services/commands/projects.ts
@@ -5,7 +5,6 @@ import * as vscode from 'vscode';
 import { ProjectTreeItem } from '../@types/projects';
 
 import * as commands from '../common/commands';
-import * as settings from '../common/settings';
 
 import { DiffPanel } from '../panel/DiffPanel';
 
@@ -25,14 +24,13 @@ export function activate (context:vscode.ExtensionContext) {
 		
 		'l13Diff.action.projects.compareWithWorkspace': async ({ project }:ProjectTreeItem) => {
 			
-			const compare = settings.get('openFavoriteAndCompare', false);
 			const workspaces = workspaceFoldersQuickPickItems();
 			
 			if (workspaces.length > 1) {
 				const value:any = await vscode.window.showQuickPick(workspaces);
-				if (value) DiffPanel.createOrShow(context, [{ fsPath: value.description }, { fsPath: project.path }], compare);
+				if (value) DiffPanel.createOrShow(context, [{ fsPath: value.description }, { fsPath: project.path }], true);
 			} else if (workspaces.length === 1) {
-				DiffPanel.createOrShow(context, [{ fsPath: workspaces[0].description }, { fsPath: project.path }], compare);
+				DiffPanel.createOrShow(context, [{ fsPath: workspaces[0].description }, { fsPath: project.path }], true);
 			} else vscode.window.showErrorMessage('No workspace available!');
 			
 		},